Relocate errors into the core package
