+++ v2/pkg/genruntime/core/error_response.go
@@ -3,16 +3,7 @@ Copyright (c) Microsoft Corporation.
 Licensed under the MIT license.
 */
 
-package genericarmclient
-
-// ErrorAdditionalInfo - The resource management error additional info.
-type ErrorAdditionalInfo struct {
-	// READ-ONLY; The additional info.
-	Info map[string]interface{} `json:"info,omitempty" azure:"ro"`
-
-	// READ-ONLY; The additional info type.
-	Type *string `json:"type,omitempty" azure:"ro"`
-}
+package core
 
 // ErrorResponse - Common error response for all Azure Resource Manager APIs to return error details for failed operations.
 type ErrorResponse struct {
